refactor(cli): split main.rs into cli/run/run_helpers submodules
Split the 3,526-line main.rs into crate-root submodules, mirroring the
spec/chumsky_{adapter,parser} and adapt/pinocchio_profile splits. Move-only:
no behavior or generated-output changes (1006 tests + all four snapshot
suites byte-identical).

Unlike the directory-module splits, main.rs is the binary crate root, so the
module tree (`mod ...`), the `crate::<module>` re-export hub, and `fn main`
must stay at the root. Everything else moves to flat crate-root submodules:

- cli.rs         clap argument definitions (Cli/Target/RuntimeOverride/
                 Commands/AristotleCommands)
- run.rs         command_name_of + the dispatch match
- run_helpers.rs dispatch-support glue (git guards, crucible bridge, CI
                 template, lint formatting, probe runners) + their tests
- main.rs        3,526 -> 71 lines: module tree, re-export hub, fn main

The cli types are re-exported from the crate root (pub(crate) use cli::{...})
so crate::Target keeps resolving for verify::regen_drift and
codegen_shared::guards. command_name_of/dispatch and the 11 helpers are bumped
to pub(crate) for cross-submodule reach; the Cli.command struct field too
(main reads it). include_str! paths are unchanged — the new files sit at the
same src/ depth as main.rs.
